Sample for v1.1.0 for Online API usage - Finalising and bug removal

+++ flask_server.py
@@ -35,26 +35,31 @@ def upload_file():
         # Get the files from the request
         config_file = request.files['Database.config']
         api_file = request.files['API.json']
+        csv_file = request.files['Test.csv']
 
         # Validate file size to ensure they are not empty
-        if config_file.filename != '' and api_file.filename != '':
+        if config_file.filename != '' and api_file.filename != '' and csv_file.filename != '':
             # Define the path for saving the files in the current working directory
             base_path = os.path.dirname(os.path.abspath(__file__))
             config_path = os.path.join(base_path, 'Database.config')
             api_path = os.path.join(base_path, 'API.json')
+            csv_file_path = os.path.join(base_path, 'Test.csv')
 
             # Replace any existing files
             if os.path.exists(config_path):
                 os.remove(config_path)
             if os.path.exists(api_path):
                 os.remove(api_path)
+            if os.path.exists(csv_file_path):
+                os.remove(csv_file_path)
 
             # Save the new files
             config_file.save(config_path)
             api_file.save(api_path)
+            csv_file.save(csv_file_path)
 
             # Return an HTML success message
-            if os.path.exists('Database.config') and os.path.exists('API.json'):
+            if os.path.exists('Database.config') and os.path.exists('API.json') and os.path.exists('Test.csv'):
                 message = database_thread()
                 if message.startswith('LIST'):
                     # Removing 'LIST' from the beginning of the message
@@ -76,12 +81,15 @@ def upload_file():
                         return f"<html><body><h1>Error</h1><h2>Error Number: 501</h2><p>Unknown error - Not Defined</p></body></html>", 501
                 else:
                     return f"<html><body><h1>Success</h1>{message}</body></html>", 200
+            else:
+                # Return an HTML error message with an error number
+                return f"<html><body><h1>Error</h1><h2>Error Number: 404</h2><p>Database.config, Test.csv and API.json files are required and cannot be empty.</p><p>{err_codes[404]}</p></body></html>", 404
         else:
             # Return an HTML error message with an error number
-            return f"<html><body><h1>Error</h1><h2>Error Number: 400</h2><p>Both Database.config and API.json files are required and cannot be empty.</p><p>{err_codes[400]}</p></body></html>", 400
+            return f"<html><body><h1>Error</h1><h2>Error Number: 404</h2><p>Database.config, Test.csv and API.json files are required and cannot be empty.</p><p>{err_codes[404]}</p></body></html>", 404
     else:
         # Return an HTML error message with an error number
-        return f"<html><body><h1>Error</h1><h2>Error Number: 400</h2><p>Both Database.config and API.json files are required.</p><p>{err_codes[400]}</p></body></html>", 400
+        return f"<html><body><h1>Error</h1><h2>Error Number: 404</h2><p>Database.config, Test.csv and API.json files are required and cannot be empty.</p><p>{err_codes[404]}</p></body></html>", 404
 
 
 if __name__ == '__main__':
